Adding Main files to Git.
I designed and developed a sophisticated website for a media agency, leveraging my expertise in HTML, CSS, and JavaScript. The project involved creating a visually stunning and user-friendly interface that effectively showcased the agency's services and portfolio. Using HTML for the structural layout, CSS for stylish and responsive design, and JavaScript for interactive elements, I ensured the website provided a seamless and engaging experience across all devices. The result is a modern, high-performance website that enhances the agency’s online visibility and allows clients to easily explore and interact with their offerings.